netlify/functions/requirements.txt
fastapi
httpx
pydantic
pydantic-settings
PyPDF2
python-docx
python-multipart